LayoutTests/fast/forms/week-multiple-fields/week-multiple-fields-fallback-format.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<script>
description('Checks incomplete datetime formats are rejected.');

var input = document.createElement('input');
input.type = 'week';
document.body.appendChild(input);

function setDateTimeFormat(pattern) {
    var value = input.value;
    getElementByPseudoId(internals.youngestShadowRoot(input), '-webkit-datetime-edit').setAttribute('pattern', pattern);
    input.value = ''; // Updates the element for new format
    input.value = value;
}

input.value = '1999-W09';

debug('Valid format');
setDateTimeFormat("'Week'ww yyyy");
shouldBeEqualToString('getUserAgentShadowTextContent(input)', 'Week09 1999');

debug("Invalid format, fallback to yyyy-'W'ww");
setDateTimeFormat('ww');
shouldBeEqualToString('getUserAgentShadowTextContent(input)', '1999-W09');
setDateTimeFormat('yyyy');
shouldBeEqualToString('getUserAgentShadowTextContent(input)', '1999-W09');

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
