third_party/blink/web_tests/fast/forms/week-multiple-fields/week-multiple-fields-fallback-format-expected.txt

Checks incomplete datetime formats are rejected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Valid format
PASS getUserAgentShadowTextContent(input) is "Week09 1999"
Invalid format, fallback to yyyy-'W'ww
PASS getUserAgentShadowTextContent(input) is "1999-W09"
PASS getUserAgentShadowTextContent(input) is "1999-W09"
PASS successfullyParsed is true

TEST COMPLETE
